pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "regex-engine"
version = "0.1.0"
description = "Ingredient regex engine"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "pydantic==2.12.2",
    "openai-agents>=0.2.0",
    "python-dotenv==1.2.2",
    "morfeusz2>=1.99",
    "pandas",
    "psycopg[binary]",
    "psycopg2-binary"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]